Time: 1 ms (99.80%) | Memory: 77.1 MB (91.83%) - LeetSync

+++ 53-maximum-subarray/maximum-subarray.java
@@ -1,6 +1,6 @@
 class Solution {
     public int maxSubArray(int[] nums) {
-        long max = Long.MIN_VALUE, sum = 0;
+        int max = Integer.MIN_VALUE, sum = 0;
 
         for (int i = 0; i < nums.length; i++) {
             sum += nums[i];
@@ -14,6 +14,6 @@ class Solution {
             }
         }
 
-        return (int) max;
+        return max;
     }
 }
